completed per event concesus

+++ pkg/client/topic.go
@@ -80,7 +80,7 @@ func ValidateTopicPayload(payload entities.ClientPayload, authState *models.Auth
 		return nil, nil, apperror.Forbidden("Subnet is required")
 	}
 	
-	if payload.EventType == uint16(constants.CreateTopicEvent) {
+	if payload.EventType ==constants.CreateTopicEvent {
 		// topic, _ := query.GetTopic(models.TopicState{
 		// 	Topic: entities.Topic{Ref: payloadData.Ref, Subnet: payloadData.Subnet},
 		// })
@@ -94,7 +94,7 @@ func ValidateTopicPayload(payload entities.ClientPayload, authState *models.Auth
 	}
 
 	payload.Data = payloadData
-	if payload.EventType == uint16(constants.CreateTopicEvent) {
+	if payload.EventType == constants.CreateTopicEvent {
 		// dont worry validating the AuthHash for Authorization requests
 		if uint64(payloadData.Timestamp) > uint64(time.Now().UnixMilli())+15000 {
 			return nil, nil, errors.New("Authorization timestamp exceeded")
